Moved chunk initialization to separated mod
