+++ crates/world_server/src/set/mod.rs
@@ -1,5 +1,7 @@
+mod chunk_initialization;
 mod chunk_management;
 mod landscape;
 
-pub(crate) use chunk_management::*;
+pub use chunk_initialization::*;
+pub use chunk_management::*;
 pub use landscape::*;
